# Conflicts:
#	src/literalizer/_language.py
#	src/literalizer/languages/ada.py
#	src/literalizer/languages/bash.py
#	src/literalizer/languages/c.py
#	src/literalizer/languages/clojure.py
#	src/literalizer/languages/cobol.py
#	src/literalizer/languages/common_lisp.py
#	src/literalizer/languages/cpp.py
#	src/literalizer/languages/crystal.py
#	src/literalizer/languages/csharp.py
#	src/literalizer/languages/d.py
#	src/literalizer/languages/dart.py
#	src/literalizer/languages/dhall.py
#	src/literalizer/languages/elixir.py
#	src/literalizer/languages/elm.py
#	src/literalizer/languages/erlang.py
#	src/literalizer/languages/forth.py
#	src/literalizer/languages/fortran.py
#	src/literalizer/languages/fsharp.py
#	src/literalizer/languages/gleam.py
#	src/literalizer/languages/go.py
#	src/literalizer/languages/groovy.py
#	src/literalizer/languages/haskell.py
#	src/literalizer/languages/hcl.py
#	src/literalizer/languages/java.py
#	src/literalizer/languages/javascript.py
#	src/literalizer/languages/json5.py
#	src/literalizer/languages/jsonnet.py
#	src/literalizer/languages/julia.py
#	src/literalizer/languages/kotlin.py
#	src/literalizer/languages/lua.py
#	src/literalizer/languages/matlab.py
#	src/literalizer/languages/mojo.py
#	src/literalizer/languages/nim.py
#	src/literalizer/languages/nix.py
#	src/literalizer/languages/norg.py
#	src/literalizer/languages/objective_c.py
#	src/literalizer/languages/ocaml.py
#	src/literalizer/languages/occam.py
#	src/literalizer/languages/odin.py
#	src/literalizer/languages/perl.py
#	src/literalizer/languages/php.py
#	src/literalizer/languages/powershell.py
#	src/literalizer/languages/purescript.py
#	src/literalizer/languages/python.py
#	src/literalizer/languages/r.py
#	src/literalizer/languages/racket.py
#	src/literalizer/languages/raku.py
#	src/literalizer/languages/ruby.py
#	src/literalizer/languages/rust.py
#	src/literalizer/languages/scala.py
#	src/literalizer/languages/scheme.py
#	src/literalizer/languages/sml.py
#	src/literalizer/languages/swift.py
#	src/literalizer/languages/systemverilog.py
#	src/literalizer/languages/tcl.py
#	src/literalizer/languages/toml.py
#	src/literalizer/languages/typescript.py
#	src/literalizer/languages/v.py
#	src/literalizer/languages/vb.py
#	src/literalizer/languages/wren.py
#	src/literalizer/languages/yaml.py
#	src/literalizer/languages/zig.py
#	tests/test_languages.py

+++ tests/integration/cases/float_special_values/Fortran_combined.f90
@@ -1,5 +1,6 @@
 module fval_m
   use, intrinsic :: ieee_arithmetic
+  use, intrinsic :: iso_fortran_env, only: int64
   implicit none
   type :: fval_t
     integer :: t = 0
@@ -7,7 +8,7 @@ module fval_m
 contains
   function fnull() result(v); type(fval_t) :: v; end function
   function fbool(b) result(v); logical, intent(in) :: b; type(fval_t) :: v; end function
-  function fint(n) result(v); integer, intent(in) :: n; type(fval_t) :: v; end function
+  function fint(n) result(v); integer(kind=int64), intent(in) :: n; type(fval_t) :: v; end function
   function freal(x) result(v); real, intent(in) :: x; type(fval_t) :: v; end function
   function fstr(s) result(v); character(len=*), intent(in) :: s; type(fval_t) :: v; end function
   function flist(a) result(v); type(fval_t), intent(in) :: a(:); type(fval_t) :: v; end function
